requirements.txt
asyncio
PyYAML
psycopg2
sqlalchemy
jupyter
nbconvert
jupyter_client
jupyter_core
pandas
